# swagger.ru.yaml
openapi: 3.0.0
info:
  title: TigerMeter Cloud API
  description: API для привязки (claim) устройств и управления универсальными инструкциями отображения с настраиваемыми текстовыми полями (symbol, topLine, mainText, bottomLine), размерами шрифтов, выравниванием и LED-управлением.
  version: 2.0.0
  x-language: ru
servers:
  - url: https://api.tigermeter.cloud
    description: Prod
  - url: http://localhost:3000
    description: Local
security:
  - bearerAuth: []
...
paths:
  /api/device-claims:
    post:
      tags: [Device]
      summary: Запросить код привязки
      description: Устройство запрашивает 6-значный код привязки (действует 5 минут). Защищено HMAC на стороне устройства.
      operationId: issueClaimCode
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimCodeIssueRequest'
      responses:
        '201':
          description: Код привязки выдан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimCodeIssueResponse'
        '404': { description: Устройство не найдено или не предварительно зарегистрировано (awaiting_claim) }
        '409': { description: Устройство не в статусе awaiting_claim }
        '500': { description: Внутренняя ошибка сервера }
...
  /api/device-claims/{code}/attach:
    post:
      tags: [Portal]
      summary: Привязать код устройства
      description: |
        Портал связывает ожидающий код привязки с текущим пользователем.
        Секрет устройства НА ЭТОМ ЭТАПЕ НЕ ГЕНЕРИРУЕТСЯ И НЕ ВОЗВРАЩАЕТСЯ.
        Он будет сгенерирован лениво при первом успешном опросе `/api/device-claims/{code}/poll` самим устройством.
      operationId: attachClaimCode
      security:
        - bearerAuth: []
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
          description: 6-значный код привязки
      responses:
        '200':
          description: Код успешно привязан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceAttachResponse'
        '400': { description: Недействительный или просроченный код }
        '401': { description: Не авторизован }
        '403': { description: Запрещено }
        '409': { description: Уже привязан }
        '500': { description: Внутренняя ошибка сервера }
  /api/device-claims/{code}/poll:
    get:
      tags: [Device]
      summary: Опрос статуса кода привязки
      description: |
        Устройство опрашивает код чтобы получить секрет после привязки пользователем.
        При первом переходе статуса в claimed секрет ГЕНЕРИРУЕТСЯ (ленивая генерация) и возвращается в ответе (одноразовая выдача).
        После успешной выдачи повторные опросы должны возвращать 404 для предотвращения повторного раскрытия секрета.
        Эндпоинт подлежит rate limiting.
      operationId: pollClaimCode
      security: []
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Привязано; секрет возвращён (одноразовый)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimPollClaimed'
        '202':
          description: Ожидание (код ещё не привязан)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending]
        '404': { description: Не найдено }
        '410': { description: Истёк срок действия }
        '500': { description: Внутренняя ошибка сервера }
  /api/devices:
    get:
      tags: [Portal]
      summary: Список моих устройств
      operationId: listMyDevices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Устройства
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceSummary'
        '401': { description: Не авторизован }
  /api/devices/{id}:
    get:
      tags: [Portal]
      summary: Детали устройства
      operationId: getDevice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Подробности устройства
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDetail'
        '401': { description: Не авторизован }
        '403': { description: Запрещено }
        '404': { description: Не найдено }
  /api/devices/{id}/revoke:
    post:
      tags: [Portal]
      summary: Отвязать (revoke) устройство
      operationId: revokeDevice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Устройство отвязано
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: revoked }
        '401': { description: Не авторизован }
        '403': { description: Запрещено }
        '404': { description: Не найдено }
...
  /api/admin/devices:
    get:
      tags: [Admin]
      summary: "Админ: список устройств"
      operationId: adminListDevices
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: string
          description: Фильтр по пользователю
        - name: status
          in: query
          schema:
            type: string
            enum: [awaiting_claim, active, revoked]
          description: Фильтр по статусу
        - name: lastSeenBefore
          in: query
          schema:
            type: string
            format: date-time
          description: Последний онлайн до
        - name: lastSeenAfter
          in: query
          schema:
            type: string
            format: date-time
          description: Последний онлайн после
      responses:
        '200':
          description: Список устройств
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceAdmin'
        '401': { description: Не авторизован }
        '403': { description: Запрещено }
  /api/admin/devices/{id}/revoke:
    post:
      tags: [Admin]
      summary: "Админ: принудительно отвязать устройство"
      operationId: adminRevoke
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200': { description: Отвязано }
        '401': { description: Не авторизован }
        '403': { description: Запрещено }
        '404': { description: Не найдено }
  /api/admin/device-claims/{code}:
    get:
      tags: [Admin]
      summary: "Админ: проверить код привязки"
      operationId: adminGetClaim
      security:
        - bearerAuth: []
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Информация о коде
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimCodeStatus'
        '404': { description: Не найдено }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceSecretAuth:
      type: http
      scheme: bearer
      bearerFormat: Device Secret
      description: Секретный токен устройства
  schemas:
...
    DeviceSummary:
      type: object
      description: Краткая информация об устройстве.
      properties:
        id: { type: string, description: Идентификатор устройства }
        name: { type: string, description: Пользовательское имя }
        status: { type: string, enum: [awaiting_claim, active, revoked], description: Статус устройства }
        lastSeen: { type: string, format: date-time, nullable: true, description: Последний хартбит }
    DeviceDetail:
      allOf:
        - $ref: '#/components/schemas/DeviceSummary'
        - type: object
          description: Детализированная информация устройства.
          properties:
            mac: { type: string, description: MAC адрес }
            userId: { type: string, nullable: true, description: Пользователь-владелец }
            battery: { type: integer, nullable: true, description: Заряд батареи (проценты) }
            led: { type: string, enum: [off, low, middle, high], nullable: true, description: Текущая яркость LED }
            secretExpiresAt: { type: string, format: date-time, nullable: true, description: Истечение текущего секрета }
            displayHash: { type: string, description: Текущий хеш инструкции }
    DeviceAdmin:
      allOf:
        - $ref: '#/components/schemas/DeviceDetail'
        - type: object
          description: Административные поля.
          properties:
            deviceSecretHash: { type: string, description: Хеш секрета }
            createdAt: { type: string, format: date-time, description: Когда создано }
...
    ClaimCodeIssueRequest:
      type: object
      required: [mac, firmwareVersion]
      description: Запрос на выдачу кода привязки.
      properties:
        mac: { type: string, description: MAC адрес устройства }
        firmwareVersion: { type: string, description: Версия прошивки }
        ip: { type: string, description: IP адрес (опционально) }
        hmac: { type: string, description: HMAC подпись против подмены }
    ClaimCodeIssueResponse:
      type: object
      description: Ответ с кодом привязки.
      properties:
        code: { type: string, description: 6-значный код }
        expiresAt: { type: string, format: date-time, description: Время истечения }
    ClaimCodeStatus:
      type: object
      description: Статус кода привязки.
      properties:
        code: { type: string, description: Код }
        status: { type: string, enum: [pending, claimed, expired], description: Текущий статус }
        deviceId: { type: string, description: Идентификатор устройства (если привязано) }
        mac: { type: string, description: MAC адрес }
        expiresAt: { type: string, format: date-time, description: Истечение }
    DeviceAttachResponse:
      type: object
      description: Результат привязки кода.
      properties:
        deviceId: { type: string, description: Идентификатор устройства }
        message: { type: string, description: Сообщение }
    ClaimPollClaimed:
      type: object
      description: Одноразовый ответ при успешной привязке. Содержит секрет устройства, который генерируется лениво в момент первого успешного опроса. Повторные опросы после получения секрета возвращают 404.
      properties:
        deviceId: { type: string, description: Идентификатор устройства }
        deviceSecret: { type: string, description: Новый секрет устройства (plaintext; показывается только один раз) }
        displayHash: { type: string, description: Текущий хеш инструкции (может быть пустым) }
        expiresAt: { type: string, format: date-time, description: Истечение секрета }
...